Move AdminController to UserController. Fix urls

+++ src/main/java/com/kabir/mydemoproject/service/UserServiceImpl.java
@@ -6,11 +6,11 @@ import com.kabir.mydemoproject.exception.ErrorMessage;
 import com.kabir.mydemoproject.exception.ResourceNotFoundException;
 import com.kabir.mydemoproject.models.ERole;
 import com.kabir.mydemoproject.models.Role;
+import com.kabir.mydemoproject.models.Seat;
 import com.kabir.mydemoproject.models.User;
 import com.kabir.mydemoproject.repository.RoleRepository;
 import com.kabir.mydemoproject.repository.UserRepository;
 import com.kabir.mydemoproject.security.jwt.JwtUtils;
-import com.kabir.mydemoproject.security.services.UserDetailsImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.authentication.AuthenticationManager;
@@ -25,10 +25,13 @@ import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
-public class UserServiceImpl implements UserService{
+public class UserServiceImpl implements UserService {
 
     @Autowired
     private UserRepository userRepository;
+
+    @Autowired
+    private SeatService seatService;
 
     @Autowired
     private AuthenticationManager authenticationManager;
@@ -64,16 +67,18 @@ public class UserServiceImpl implements UserService{
 
     @Override
     public List<User> getUsers() {
-        List<User> userList = new ArrayList<>();
-        userRepository.findAll().forEach(userList::add);
-        return userList;
+        return new ArrayList<>(userRepository.findAll());
     }
 
     @Override
     public User getUser(Long id) {
-        Authentication authentication=SecurityContextHolder.getContext().getAuthentication();
-        Long idFromJwt= Long.parseLong(String.valueOf(authentication.getPrincipal()));
-        if(!idFromJwt.equals(id)){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Long idFromJwt = Long.parseLong(String.valueOf(authentication.getPrincipal()));
+        List<String> roles = authentication.getAuthorities().stream()
+                .map(GrantedAuthority::getAuthority)
+                .collect(Collectors.toList());
+
+        if (!idFromJwt.equals(id)&&!roles.contains("ROLE_ADMIN")) {
             throw new AccessDeniedException("You are not supposed to do this.");
         }
         Optional<User> optionalUser = userRepository.findById(id);
@@ -103,7 +108,7 @@ public class UserServiceImpl implements UserService{
         user.setPassword(encoder.encode(userDto.getPassword()));
         Set<Role> roles = new HashSet<>();
         Optional<Role> userRole = roleRepository.findByName(ERole.ROLE_USER);
-        if(!userRole.isPresent()){
+        if (!userRole.isPresent()) {
             throw new ResourceNotFoundException(new ErrorMessage("No Role Found"));
         }
         roles.add(userRole.get());
@@ -115,10 +120,19 @@ public class UserServiceImpl implements UserService{
 
     @Override
     public MyResponse deleteUser(Long id) {
-        Authentication authentication=SecurityContextHolder.getContext().getAuthentication();
-        Long idFromJwt= Long.parseLong(String.valueOf(authentication.getPrincipal()));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Long idFromJwt = Long.parseLong(String.valueOf(authentication.getPrincipal()));
+        List<String> roles = authentication.getAuthorities().stream()
+                .map(GrantedAuthority::getAuthority)
+                .collect(Collectors.toList());
+
+        if (!idFromJwt.equals(id)&&!roles.contains("ROLE_ADMIN")) {
+            throw new AccessDeniedException("You are not supposed to do this.");
+        }
+
+
         Optional<User> optionalUser = userRepository.findById(id);
-        if(!idFromJwt.equals(id)||!optionalUser.isPresent()){
+        if (!idFromJwt.equals(id) || !optionalUser.isPresent()) {
             throw new AccessDeniedException("You are not supposed to do this.");
         }
         userRepository.deleteById(id);
@@ -126,10 +140,14 @@ public class UserServiceImpl implements UserService{
     }
 
     @Override
-    public User updateUser(Long id,Password password) {
-        Authentication authentication=SecurityContextHolder.getContext().getAuthentication();
-        Long idFromJwt= Long.parseLong(String.valueOf(authentication.getPrincipal()));
-        if(!idFromJwt.equals(id)){
+    public User updateUser(Long id, Password password) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Long idFromJwt = Long.parseLong(String.valueOf(authentication.getPrincipal()));
+        List<String> roles = authentication.getAuthorities().stream()
+                .map(GrantedAuthority::getAuthority)
+                .collect(Collectors.toList());
+
+        if (!idFromJwt.equals(id)&&!roles.contains("ROLE_ADMIN")) {
             throw new AccessDeniedException("You are not supposed to do this.");
         }
         Optional<User> optionalUser = userRepository.findById(id);
@@ -149,11 +167,33 @@ public class UserServiceImpl implements UserService{
         SecurityContextHolder.getContext().setAuthentication(authentication);
         String jwt = jwtUtils.generateJwtToken(authentication);
 
-        UserDetailsImpl userDetails = (UserDetailsImpl) authentication.getPrincipal();
-        List<String> roles = userDetails.getAuthorities()
-                .stream()
-                .map(GrantedAuthority::getAuthority)
-                .collect(Collectors.toList());
         return new JwtResponse(jwt, "Bearer");
+    }
+
+    @Override
+    public List<Seat> createHall(int numberOfSeats) {
+        for (int i = 0; i < numberOfSeats; i++) {
+            Seat seat = new Seat();
+            seat.setBooked(false);
+            seatService.create(seat);
+            for(int j=0;j<100000;j++){
+                System.out.println();
+            }
+        }
+        return seatService.getAllSeats();
+    }
+
+    @Override
+    public MyResponse makeAdmin(Long id) {
+        User user = userRepository.getById(id);
+        Optional<Role> userRole = roleRepository.findByName(ERole.ROLE_ADMIN);
+        if(!userRole.isPresent()){
+            throw new ResourceNotFoundException(new ErrorMessage("No Role Found"));
+        }
+        Set<Role> userRoles=user.getRoles();
+        userRoles.add(userRole.get());
+        user.setRoles(userRoles);
+        userRepository.save(user);
+        return new MyResponse(user.getUsername()+" has been made an admin.");
     }
 }
